pyproject.toml
[project]
name = "fund-tracker"
version = "0.1.0"
description = "Project to emulate a portfolio"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.11",
    "finnhub-python>=2.4.22",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "router>=0.1",
    "typing>=3.10.0.0",
    "uvicorn>=0.34.0",
    "yfinance>=0.2.54",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0,<8.0.0",
    "ruff>=0.9.9",
]

[tools.uv]
default-groups = []

[tools.uv.dev-dependencies]
test=["pytest>=7.0.0,<8.0.0"]
